Fix bug and update Readme

+++ vietocr/model/seqmodel/transformer.py
@@ -42,7 +42,7 @@ class LanguageTransformer(nn.Module):
         tgt = self.pos_enc(self.embed_tgt(tgt) * math.sqrt(self.d_model))
         
         output = self.transformer(src, tgt, tgt_mask=tgt_mask, src_key_padding_mask=src_key_padding_mask,
-                                  tgt_key_padding_mask=tgt_key_padding_mask, memory_key_padding_mask=memory_key_padding_mask)
+                                  tgt_key_padding_mask=tgt_key_padding_mask.float(), memory_key_padding_mask=memory_key_padding_mask)
 #        output = rearrange(output, 't n e -> n t e')
         output = output.transpose(0, 1)
         return self.fc(output)
